requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.0
django-allauth>=0.54.0
celery>=5.3.0
redis>=5.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
dj-database-url>=2.1.0
drf-spectacular>=0.27.0
requests>=2.31.0
beautifulsoup4>=4.12.0